+++ packages/debug-apps/react-ssr/src/routes.ts
@@ -1,0 +1,12 @@
+import {remixRoutesOptionAdapter} from "@react-router/remix-routes-option-adapter"
+
+import {hybridRoutes} from "@qualcomm-ui/react-router-utils/node"
+
+export const routes = remixRoutesOptionAdapter((defineRoutes) => {
+  return hybridRoutes("routes", defineRoutes, {
+    appDir: "src",
+    ignoredRouteFiles: ["**/*components/**/*", "**/*demos/**/*"],
+  })
+})
+
+export default routes