+++ scripts/lint/types.ts
@@ -1,0 +1,4 @@
+export interface PackageTiming {
+  duration: number
+  package: string
+}