+++ packages/common/core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @qualcomm-ui/core
 
+## 1.0.5 (2025/11/13)
+
+### Bug Fixes
+
+- add stub ts files for npm ts badge
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/dom, @qualcomm-ui/utils]
+
+## 1.0.4 (2025/11/13)
+
+### Miscellaneous Chores
+
+- provide empty default type for npm typescript badge
+
 ## 1.0.3 (2025/11/13)
 
 ### Bug Fixes