+++ packages/common/dom/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @qualcomm-ui/dom
 
+## 1.0.4 (2025/11/14)
+
+### Bug Fixes
+
+- ssr guard on raf call
+
+## 1.0.3 (2025/11/13)
+
+### Bug Fixes
+
+- add stub ts files for npm ts badge
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/utils]
+
 ## 1.0.2 (2025/11/12)
 
 ### Bug Fixes