+++ packages/frameworks/react-mdx/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @qualcomm-ui/react-mdx
 
+## 1.0.4 (2025/11/14)
+
+### Bug Fixes
+
+- element preflight styles
+- shiki inner html padding
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/react]
+
+## 1.0.3 (2025/11/13)
+
+### Bug Fixes
+
+- add stub ts files for npm ts badge
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/core, @qualcomm-ui/mdx-common, @qualcomm-ui/utils, @qualcomm-ui/react, @qualcomm-ui/react-core]
+
 ## 1.0.2 (2025/11/12)
 
 ### Bug Fixes