+++ packages/frameworks/react-router-utils/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @qualcomm-ui/react-router-utils
 
+## 1.0.4 (2025/11/14)
+
+### Performance Improvements
+
+- use simple css transition on mobile
+
+### Bug Fixes
+
+- exclude dependencies from bundle
+- shiki inner html padding
+
+### Miscellaneous Chores
+
+- add my-ua-parser dependency
+
+## 1.0.3 (2025/11/13)
+
+### Bug Fixes
+
+- add stub ts files for npm ts badge
+
 ## 1.0.2 (2025/11/12)
 
 ### Bug Fixes